Document and isolate side panel title syncing
